Generate empty findlib.conf
The path that was previously set was incorrect but this was still
working as dune is properly setting findlib env variables for it to find
packages.
We still need to provide an initial configuration file as the dune port
for findlib doesn't come with one and therefore triggers a crash when
the library is loaded. We can just point it to an empty conf and let
dune do the rest.