Add a --locks option to mdx rule
This can be used to set locks on generated dune aliases and to prevent
hasardous concurrent execution of `mdx test` commands.

It might be worth improving this by setting locks on some of the
dependencies to prevent, for example, concurrent writing of `mdx test`
processes to the same `.corrected` file or concurrent execution of shell
snippets in the same directories.